feat: subject groups
Add a subject group model to the Django app and the API. Subject groups
have a list of subjects and a list of protocols associated with those
subjects. Datasets have a list of subject groups in `dataset.groups`.

Refactor the Data and Trial Design views to use `dataset.groups` for
tabbed views.