feat: download data as a CSV
Convert group dosing data and group observations to a single CSV file for all subject groups. CSV headers are taken from a standardised list used by the Django app. Remove the previous download button, on the last step of the data stepper, and replace it with a download button in the Data tab.