feat: project groups (#414)
Subject groups that belong to a project but not a dataset. These are created by the New Group tab under Trial Design. A project group has protocols, but no subjects. It can generate simulations but it isn't listed under the data tab. Project groups are listed with dataset groups in the Simulations tab, and project simulations are plotted alongside dataset simulations by default.