refactor(frontend) align data upload dosing with Trial Design  (#639)
When uploading data without any dosing rows, use the project dosing protocols to determine the default dosing compartments. Previously, an uploaded subject group could only be mapped to a single dosing compartment, even if the model allowed for both IV and SC dosing. With the changes here, subject groups can be dosed from each compartment linked to the model (when an uploaded CSV doesn't have any dosing rows.)

Alter the dosing check in the sidebar so that it only flags up a group if none of that group's protocols have a dose set, rather than checking if any doses are 0.